requirements.txt
amqp==5.2.0
astroid==3.2.2
billiard==4.2.0
black==24.4.2
blinker==1.8.2
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.4
dill==0.3.8
dnspython==2.6.1
dominate==2.9.1
flake8==7.1.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pyflakes==3.2.0
pylint==3.2.5
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
reportlab==4.2.2
six==1.16.0
tomlkit==0.12.5
tzdata==2024.1
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.0.3
